# openapi/openapi.json
{
  "components": {
...
    "securitySchemes": {
      "bearerAuth": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Technical foundation of the Barber Platform core API.",
    "title": "Barber Core API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/professionals/{professionalId}/activate": {
      "post": {
        "description": "Allowed roles: admin, manager.",
        "operationId": "activateProfessional",
        "parameters": [
          {
            "in": "path",
            "name": "professionalId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "bio": {
                      "maxLength": 1000,
                      "minLength": 1,
                      "type": "string"
                    },
                    "createdAt": {
                      "format": "date-time",
                      "type": "string"
                    },
                    "id": {
                      "format": "uuid",
                      "type": "string"
                    },
                    "name": {
                      "maxLength": 120,
                      "minLength": 2,
                      "type": "string"
                    },
                    "status": {
                      "anyOf": [
                        {
                          "enum": [
                            "active"
                          ],
                          "type": "string"
                        },
                        {
                          "enum": [
                            "inactive"
                          ],
                          "type": "string"
                        }
                      ]
                    },
                    "updatedAt": {
                      "format": "date-time",
                      "type": "string"
                    }
                  },
                  "required": [
                    "id",
                    "name",
                    "status",
                    "createdAt",
                    "updatedAt"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Default Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "detail": {
                      "type": "string"
                    },
                    "errors": {
                      "items": {
                        "properties": {
                          "code": {
                            "type": "string"
                          },
                          "field": {
                            "type": "string"
                          },
                          "message": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "field",
                          "message"
                        ],
                        "type": "object"
                      },
                      "type": "array"
                    },
                    "instance": {
                      "type": "string"
                    },
                    "requestId": {
                      "format": "uuid",
                      "type": "string"
                    },
                    "status": {
                      "type": "integer"
                    },
                    "title": {
                      "type": "string"
                    },
                    "type": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status",
                    "detail",
                    "instance",
                    "code",
                    "requestId"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Default Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "detail": {
                      "type": "string"
                    },
                    "errors": {
                      "items": {
                        "properties": {
                          "code": {
                            "type": "string"
                          },
                          "field": {
                            "type": "string"
                          },
                          "message": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "field",
                          "message"
                        ],
                        "type": "object"
                      },
                      "type": "array"
                    },
                    "instance": {
                      "type": "string"
                    },
                    "requestId": {
                      "format": "uuid",
                      "type": "string"
                    },
                    "status": {
                      "type": "integer"
                    },
                    "title": {
                      "type": "string"
                    },
                    "type": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status",
                    "detail",
                    "instance",
                    "code",
                    "requestId"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Default Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "detail": {
                      "type": "string"
                    },
                    "errors": {
                      "items": {
                        "properties": {
                          "code": {
                            "type": "string"
                          },
                          "field": {
                            "type": "string"
                          },
                          "message": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "field",
                          "message"
                        ],
                        "type": "object"
                      },
                      "type": "array"
                    },
                    "instance": {
                      "type": "string"
                    },
                    "requestId": {
                      "format": "uuid",
                      "type": "string"
                    },
                    "status": {
                      "type": "integer"
                    },
                    "title": {
                      "type": "string"
                    },
                    "type": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status",
                    "detail",
                    "instance",
                    "code",
                    "requestId"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Default Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "detail": {
                      "type": "string"
                    },
                    "errors": {
                      "items": {
                        "properties": {
                          "code": {
                            "type": "string"
                          },
                          "field": {
                            "type": "string"
                          },
                          "message": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "field",
                          "message"
                        ],
                        "type": "object"
                      },
                      "type": "array"
                    },
                    "instance": {
                      "type": "string"
                    },
                    "requestId": {
                      "format": "uuid",
                      "type": "string"
                    },
                    "status": {
                      "type": "integer"
                    },
                    "title": {
                      "type": "string"
                    },
                    "type": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status",
                    "detail",
                    "instance",
                    "code",
                    "requestId"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Default Response"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "detail": {
                      "type": "string"
                    },
                    "errors": {
                      "items": {
                        "properties": {
                          "code": {
                            "type": "string"
                          },
                          "field": {
                            "type": "string"
                          },
                          "message": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "field",
                          "message"
                        ],
                        "type": "object"
                      },
                      "type": "array"
                    },
                    "instance": {
                      "type": "string"
                    },
                    "requestId": {
                      "format": "uuid",
                      "type": "string"
                    },
                    "status": {
                      "type": "integer"
                    },
                    "title": {
                      "type": "string"
                    },
                    "type": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status",
                    "detail",
                    "instance",
                    "code",
                    "requestId"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Default Response"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Activates one professional.",
        "tags": [
          "Professionals"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "Local development server",
      "url": "http://localhost:3000"
    }
  ],
...
}